requirements.txt
appdirs==1.4.3
APScheduler==3.3.1
backports.ssl-match-hostname==3.5.0.1
funcsigs==1.0.2
futures==3.1.1
logging==0.4.9.6
packaging==16.8
PyExecJS==1.4.0
pyparsing==2.2.0
pytz==2017.2
requests==2.14.2
six==1.10.0
socketIO-client-2==0.7.2
tzlocal==1.4
websocket-client==0.40.0
